Make Registrant/Registrar use same TARA controller
